补充资料

+++ openglTest/myMat.h
@@ -1,17 +1,17 @@
 #pragma once
 class vec2 {
 	public:
-		double x;
-		double y;
+		GLfloat x;
+		GLfloat y;
 
 	public:
-		vec2(double x, double y);
+		vec2(GLfloat x, GLfloat y);
 		vec2 operator +(const vec2& v2) const;
 		vec2 operator -(const vec2& v2) const;
-		vec2 operator /(const double num) const;
+		vec2 operator /(const GLfloat num) const;
 };
 
-vec2::vec2(double x=0, double y=0) {
+vec2::vec2(GLfloat x=0, GLfloat y=0) {
 	this->x = x;
 	this->y = y;
 }
@@ -26,7 +26,7 @@ inline vec2 vec2::operator-(const vec2& v2) const
 	return vec2(this->x - v2.x, this->y - v2.y);
 }
 
-inline vec2 vec2::operator/(const double num) const
+inline vec2 vec2::operator/(const GLfloat num) const
 {
 	return vec2(this->x / num, this->y / num);
 }
